requirements.txt
opencv-python
numpy
flask
Pillow
ultralytics
gradio
